#2308 generalize laziness in AccessesPostsLazily trait
